more refactoring + tests--merge test isn't running
