+++ src/atmospherics/gases/gastype.rs
